+++ Renderer/src/main/java/net/buildabrowser/babbrowser/renderer/paint/backend/LoadedImage.java
@@ -2,8 +2,8 @@ package net.buildabrowser.babbrowser.renderer.paint.backend;
 
 public interface LoadedImage {
  
-  float width();
+  int width();
 
-  float height();
+  int height();
 
 }
